Move rauc.db to boot partition
The RAUC metadata file contains information that is tightly related to
the system and kernel partitions. With the possibility to migrate data
disk, the rauc.db can contain bogus information when moved to a
different system. Removal of the file on "device wipe" is also not
desirable, because the information about slot status is lost.

Relocate the rauc.db to the boot partition on system upgrade and adjust
the script for recreating it. The downside is that its content would be
recreated if the boot slot is changed or system downgraded but this
should be handled quite gracefully.

Also remove the raucdb-first-boot service which is no longer necessary
with the file not present in the data partition.